requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
psycopg2-binary